Restructure HasReader/State modules
So that the Reader(IO)Ref and ReadState instances can be exported from
the top-level HasReader/HasState module without circular dependencies.

See https://github.com/tweag/capabilities-via/pull/17#discussion_r209925515
See https://github.com/tweag/capabilities-via/pull/19#discussion_r209890054
