+++ src/Example/CountLog.hs
@@ -32,7 +32,6 @@ import qualified Streaming.Prelude as S
 
 import HasReader
 import HasState
-import HasState.Ref
 import HasStream
 import HasWriter
