+++ src/HasState.hs
@@ -1,19 +1,3 @@
-{-# OPTIONS_GHC -Wno-simplifiable-class-constraints #-}
-{-# LANGUAGE AllowAmbiguousTypes #-}
-{-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE FunctionalDependencies #-}
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
-{-# LANGUAGE InstanceSigs #-}
-{-# LANGUAGE KindSignatures #-}
-{-# LANGUAGE MagicHash #-}
-{-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE QuantifiedConstraints #-}
-{-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE TypeApplications #-}
-{-# LANGUAGE TypeInType #-}
-{-# LANGUAGE UnboxedTuples #-}
-{-# LANGUAGE UndecidableInstances #-}
-
 module HasState
   ( HasState (..)
   , get
@@ -23,97 +7,11 @@ module HasState
   , modify'
   , gets
   , MonadState (..)
+  , ReaderIORef (..)
+  , ReaderRef (..)
   , module Accessors
   ) where
 
-import Control.Lens (set, view)
-import Control.Monad.IO.Class (MonadIO)
-import Control.Monad.Primitive (PrimMonad)
-import qualified Control.Monad.State.Class as State
-import Control.Monad.Trans.Class (MonadTrans, lift)
-import Data.Coerce (Coercible, coerce)
-import qualified Data.Generics.Product.Fields as Generic
-import GHC.Exts (Proxy#, proxy#)
-import GHC.Generics (Generic)
-
 import Accessors
-
-
-class Monad m
-  => HasState (tag :: k) (s :: *) (m :: * -> *) | tag m -> s
-  where
-    get_ :: Proxy# tag -> m s
-    put_ :: Proxy# tag -> s -> m ()
-    state_ :: Proxy# tag -> (s -> (a, s)) -> m a
-
-get :: forall tag s m. HasState tag s m => m s
-get = get_ (proxy# @_ @tag)
-
-put :: forall tag s m. HasState tag s m => s -> m ()
-put = put_ (proxy# @_ @tag)
-
-state :: forall tag s m a. HasState tag s m => (s -> (a, s)) -> m a
-state = state_ (proxy# @_ @tag)
-
-modify :: forall tag s m. HasState tag s m => (s -> s) -> m ()
-modify f = state @tag $ \s -> ((), f s)
-
-modify' :: forall tag s m. HasState tag s m => (s -> s) -> m ()
-modify' f = do
-  s' <- get @tag
-  put @tag $! f s'
-
-gets :: forall tag s m a. HasState tag s m => (s -> a) -> m a
-gets f = do
-  s <- get @tag
-  pure (f s)
-
-
--- | Derive 'HasState' from @m@'s
--- 'Control.Monad.State.Class.MonadState' instance.
-newtype MonadState (m :: * -> *) (a :: *) = MonadState (m a)
-  deriving (Functor, Applicative, Monad, MonadIO, PrimMonad)
-instance State.MonadState s m => HasState tag s (MonadState m) where
-  get_ _ = coerce @(m s) State.get
-  put_ _ = coerce @(s -> m ()) State.put
-  state_ :: forall a. Proxy# tag -> (s -> (a, s)) -> MonadState m a
-  state_ _ = coerce @((s -> (a, s)) -> m a) State.state
-
-
--- | Convert the state using safe coercion.
-instance
-  ( Coercible from to, HasState tag from m
-  , forall x y. Coercible x y => Coercible (m x) (m y) )
-  => HasState tag to (Coerce to m)
-  where
-    get_ tag = coerce @(m from) $ get_ tag
-    put_ tag = coerce @(from -> m ()) $ put_ tag
-    state_ :: forall a. Proxy# tag -> (to -> (a, to)) -> Coerce to m a
-    state_ tag = coerce @((from -> (a, from)) -> m a) $ state_ tag
-
-
--- | Zoom in on the record field @field@ of type @s@ in the state @s'@.
-instance
-  -- The constraint raises @-Wsimplifiable-class-constraints@.
-  -- This could be avoided by instead placing @HasField'@s constraints here.
-  -- Unfortunately, it uses non-exported symbols from @generic-lens@.
-  ( Generic record, Generic.HasField' field record v, HasState tag record m )
-  => HasState tag v (Field field m)
-  where
-    get_ _ = coerce @(m v) $
-      gets @tag $ view (Generic.field' @field)
-    put_ _ = coerce @(v -> m ()) $
-      modify @tag . set (Generic.field' @field @record)
-    state_ :: forall a. Proxy# tag -> (v -> (a, v)) -> Field field m a
-    state_ _ = coerce @((v -> (a, v)) -> m a) $
-      state @tag . Generic.field' @field @_ @_ @((,) a)
-
-
--- | Lift one layer in a monad transformer stack.
-instance (HasState tag s m, MonadTrans t, Monad (t m))
-  => HasState tag s (Lift (t m))
-  where
-    get_ _ = coerce $ lift @t @m $ get @tag @s
-    put_ _ = coerce $ lift @t @m . put @tag @s
-    state_ :: forall a. Proxy# tag -> (s -> (a, s)) -> Lift (t m) a
-    state_ _ = coerce $ lift @t @m . state @tag @s @m @a
+import HasState.Internal.Class
+import HasState.Internal.Instances
