+++ src/HasState/Internal/Class.hs
@@ -1,0 +1,51 @@
+{-# LANGUAGE AllowAmbiguousTypes #-}
+{-# LANGUAGE FunctionalDependencies #-}
+{-# LANGUAGE KindSignatures #-}
+{-# LANGUAGE MagicHash #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE TypeInType #-}
+
+module HasState.Internal.Class
+  ( HasState (..)
+  , get
+  , put
+  , state
+  , modify
+  , modify'
+  , gets
+  ) where
+
+import GHC.Exts (Proxy#, proxy#)
+
+
+class Monad m
+  => HasState (tag :: k) (s :: *) (m :: * -> *) | tag m -> s
+  where
+    get_ :: Proxy# tag -> m s
+    put_ :: Proxy# tag -> s -> m ()
+    state_ :: Proxy# tag -> (s -> (a, s)) -> m a
+
+get :: forall tag s m. HasState tag s m => m s
+get = get_ (proxy# @_ @tag)
+
+put :: forall tag s m. HasState tag s m => s -> m ()
+put = put_ (proxy# @_ @tag)
+
+state :: forall tag s m a. HasState tag s m => (s -> (a, s)) -> m a
+state = state_ (proxy# @_ @tag)
+
+modify :: forall tag s m. HasState tag s m => (s -> s) -> m ()
+modify f = state @tag $ \s -> ((), f s)
+
+modify' :: forall tag s m. HasState tag s m => (s -> s) -> m ()
+modify' f = do
+  s' <- get @tag
+  put @tag $! f s'
+
+gets :: forall tag s m a. HasState tag s m => (s -> a) -> m a
+gets f = do
+  s <- get @tag
+  pure (f s)
